Close #4

+++ src/Controller/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
+use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -15,9 +16,13 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks", name="task_list")
      */
-    public function list(): Response
+    public function list(TaskRepository $taskRepository): Response
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
+        $tasks = $taskRepository->findBy(['user' => $this->getUser()]);
+
+        return $this->render('task/list.html.twig', [
+            'tasks' => $tasks
+        ]);
     }
 
     /**
@@ -26,6 +31,7 @@ class TaskController extends AbstractController
     public function create(Request $request, EntityManagerInterface $em): Response
     {
         $task = new Task();
+        $task->setUser($this->getUser());
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -48,6 +54,10 @@ class TaskController extends AbstractController
      */
     public function edit(Task $task, Request $request, EntityManagerInterface $em): Response
     {
+        if ($task->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException();
+        }
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -71,6 +81,10 @@ class TaskController extends AbstractController
      */
     public function toggleTask(Task $task, EntityManagerInterface $em): Response
     {
+        if ($task->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException();
+        }
+
         $task->toggle(!$task->isDone());
         $em->flush();
 
@@ -84,6 +98,10 @@ class TaskController extends AbstractController
      */
     public function deleteTask(Task $task, EntityManagerInterface $em): Response
     {
+        if ($task->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException();
+        }
+
         $em->remove($task);
         $em->flush();
